+++ smartcontract/sdk/rs/src/commands/accesspass/mod.rs
@@ -3,3 +3,4 @@ pub mod close;
 pub mod get;
 pub mod list;
 pub mod set;
+pub mod set_feeds;